usertask Dao

+++ src/main/java/xyz/tobebetter/service/impl/ProposeService.java
@@ -1,7 +1,5 @@
 package xyz.tobebetter.service.impl;
 
-
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import xyz.tobebetter.dao.ProposeDao;
@@ -12,40 +10,86 @@ import xyz.tobebetter.util.EntityUtil;
 import xyz.tobebetter.util.MessageUtil;
 
 import java.util.List;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+import xyz.tobebetter.entity.Page;
 
 /**
- * Created by zhuleqi on 2018/2/23.
+ * Propse Created by zhuleqi on 2018/2/23.
  */
 @Service
-public class ProposeService implements ProposeServiceI {
+public class ProposeService<T extends Propose> implements ProposeServiceI<T> {
+
     @Autowired
-    private ProposeDao proposeDao;
-    public Message create(String connect, String message) {
-        Propose propose = EntityUtil.initEnity(new Propose());
-        propose.setConnect(connect);
-        propose.setMessage(message);
+    private ProposeDao<T> proposeDao;
+
+    @Override
+    public Message<T> create(T t) {
+        t = EntityUtil.initEnity(t);
+
         try {
-            proposeDao.create(propose);
-            return MessageUtil.createMessage(Message.SUCCESS,"ok",null);
-        }catch (Exception ex){
+            proposeDao.create(t);
+
+        } catch (Exception ex) {
             ex.printStackTrace();
+            return MessageUtil.createErrorMessage(ex.getMessage(), t);
 
         }
-        return MessageUtil.createMessage(Message.ERROR,"error",null);
+        return MessageUtil.createMessage(Message.SUCCESS, "ok", t);
+    }
 
+    @Override
+    public Message<T> delete(String id) {
+        T t = null;
+        try {
+            t = this.proposeDao.findById(id);
+            if (t == null) {
+                return MessageUtil.createErrorMessage("没有找到要删除的数据", id);
+            }
+            proposeDao.delete(id);
+        } catch (Exception ex) {
+            Logger.getLogger(ProposeService.class.getName()).log(Level.SEVERE, null, ex);
+            return MessageUtil.createErrorMessage(ex.getMessage(), id);
+        }
+        return this.toMessage(t);
 
     }
 
-    public Message delete(String id) {
-       proposeDao.delete(id);
-        return MessageUtil.createMessage(Message.SUCCESS,"ok",null);
-
-
+    @Override
+    public Message<T[]> findAll() {
+        List<T> proposeList = null;
+        try {
+            proposeList = proposeDao.findAll();
+        } catch (Exception ex) {
+            Logger.getLogger(ProposeService.class.getName()).log(Level.SEVERE, null, ex);
+            return MessageUtil.createErrorMessage(ex.getMessage(), null);
+        }
+        return this.toMessage(proposeList);
     }
 
-    public Message findAll() {
-        List<Propose> proposeList = proposeDao.findAll();
-        return MessageUtil.createMessage(Message.SUCCESS,"ok",proposeList);
+    @Override
+    public Message<T> findById(String id) {
+        T t = null;
+        try {
+            t = this.proposeDao.findById(id);
+        } catch (Exception ex) {
+            Logger.getLogger(ProposeService.class.getName()).log(Level.SEVERE, null, ex);
+            return MessageUtil.createErrorMessage(ex.getMessage(), null);
+        }
+
+        return this.toMessage(t);
+    }
+
+    @Override
+    public Message<T[]> find(Page page) {
+        List<T> proposeList = null;
+        try {
+            proposeList = proposeDao.find(page);
+        } catch (Exception ex) {
+            Logger.getLogger(ProposeService.class.getName()).log(Level.SEVERE, null, ex);
+            return MessageUtil.createErrorMessage(ex.getMessage(), null);
+        }
+        return this.toMessage(proposeList);
     }
 
 }
